backend/weather_assistant/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.12.5
python-dotenv==1.0.1

typer==0.15.1
rich==13.9.4
click==8.1.7

httpx==0.28.1

pytest==8.3.4
